Reworked Show_Graph() and minor ameliorations
In 'Show_Graph()', a code snippet was added, allowing to plot all periods of a time resolved spectrum in one averaged period. Additionally the name of variable psd was changed to data for uniformity. In 'in_phase_angle()' a minor problem with datasets from different spectrometers was marked (sometimes they are inverted, which leads to garbage results). Need to rework this soon.