Easy way to use rectangular function or higher sine harmonics added
Implemented a way to use different harmonics and a rectangular function for demodulation. Decision is possible by filling in a new textfield in the gui. PSD_calc now uses the textfield too to get the correct number of periods instead of the previous calculation. And some minor changes to the code. The readme got updated and overhauled to be more precise.